# apps/dashboard/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Webhook-driven rolling deployment orchestrator for Docker Compose stacks",
    "title": "RollHook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{id}/logs": {
      "get": {
        "description": "Streams deployment log lines as Server-Sent Events (`text/event-stream`). Each `data:` event carries a single log line. The stream ends with a `data: [DONE]` sentinel when the job reaches a terminal state.",
        "operationId": "stream-job-logs",
        "parameters": [
          {
            "description": "Job UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE log stream"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stream job logs",
        "tags": [
          "Jobs"
        ]
      }
    }
  }
}